Blueprint europa idea 1
As you can see, the top left corner, is the internal structure of the ship, made of a hexagonal shape divided to make it even more resistant. on the bottom left corner, are layers, first, made of aluminium-lithium because of its lightness, then tungsten, water, grid-patterned lead filled with polyethylene enriched with borum and in the middle of that layers an active water refrigerant. The middle, is the structure of energy and engines. Nuclear reactor with two water pumps, uranium pills as fuel, then, the turbine leads the steam to a heat exchanger to refrigerate it. a channel with a valve and an invar thermal sealant. After that, the steam will be rerouted to a tungsten induction chamber to make hydrogen. The hydrogen made out of that will be taken to a hydrogen tank. Then, as usual, the cryogenic engines will heat the hydrogen and oxygen to propel the space ship. at the right, its the diagram of the possible distribution, the nuclear reactor would be divided from the hydrogen tank to avoid possible contamination. Also, the reactor would be isolated